requirements-dev.txt
# Development dependencies for the Flask app
# Includes testing libraries and other tools for development

pytest
pytest-flask